modified code according your pr
